Add a new multi-port UDP multicast provider
Background:
The current UDP multicast provider manages subscriptions entirely in
userspace, which means that it receives every LCM message, and then either drops
or dispatches it.

However, in situations where there are a large number of processes on a single
core and a large number of messages flying around, these unnecessary wake ups
can cause a significant amount of overhead.

One way to reduce the overhead is to let the operating system do the filtering
by spreading the LCM traffic over a set of multicast channels (ports).

Usage:
You can enable the multi-port udp multicast provider by setting the lcm_url to:
mpudpm://239.255.76.67:7667

The default number of ports is 500. You can change this value by specifying the
nports option:
mpudpm://239.255.76.67:7667?ttl=0?nports=1000

With 1 port, the changes are backwards compatible with the original udpm
provider.

How It Works:
Instead of operating on a single UDP multicast  IP:port combo, LCM will operate
on a range of ports. To map a channel to a port, we use a string hash function.
With a sufficiently large range of ports, each channel will be mapped to a
unique port.

Unfortunately, hashing the channel names makes it more complicated to handle
subscriptions that are specified using a regex, since the subscribers won't
know which ports to listen on. We therefore must create a protocol to
disseminate the information about the set of channels being published to all
nodes in the system.

The lowest port in the port range is used as a communication channel for the
lcm nodes to communicate. Channels that begin with a '#!' are reserved for
communication between nodes.

Each node maintains a mapping of ALL channels in use by the system.

During normal operation, each node will broadcast its channel to port map
update under the following conditions:
- When it begins publishing on a new channel
- If it gets a request from another node
- If it gets an update from another node with an incomplete set of channels.
- If 5 seconds have elapsed since it last published or received an update that
  was identical to its current mapping.

Because we use UDP multicast to communicate between the nodes, we cancel a
scheduled broadcast of the channel to port mapping if a node receives an update
with an up-to-date mapping because the broadcast should have gone to all nodes.
Due to these "cancellations", during steady state (normal operation), only 1
update message will be broadcast every 5 seconds regardless of the number of
nodes in the system.

Implementation Details:
- There is a lot of shared functionality between the original udp multicast
  provider and the new multi-port one. I re-factored things to move all of this
shared code to to udpm_util.h/c
- The patch uses glib regexes for everything. Lcm now requires a glib version >=
  2.14.0"

git-svn-id: https://lcm.googlecode.com/svn/branches/multiport_udp_multicast@763 989093bb-e83e-0410-a25a-9184cbcad8d0