fix the binary search in lcm_eventlog_seek_to_timestamp
The timestamp seek functionality works by doing a binary search on raw file
data, rather than on the events in the log. The search location is represented
as a fraction of the file length. It adjusts the search fraction (named `frac`),
seeks to that point of the file, and then scans forward looking for the start
of the next event. It compares this event's timestamp with the desired, and
repeats the binary search as necessary until the timestamp is found, the
search window closes, or the search fraction delta is below a threshold.

The bug was re-adjusting the search fraction to match the start of the nearest
event. This causes the search to terminate if the same event is found in
consecutive searches, which can happen if there is an event larger than half
the window size.