fix bitshifting of signed integer types
This change fixes issues that were caused by the unspecified behavior when shifting a negative signed integer.
See Section 5.8 paragraph 3 of the standard
http://open-std.org/JTC1/SC22/WG21/docs/papers/2015/n4527.pdf

In the C/C++ lcm-gen, switch to using use uint64_t for computing the lcmtype hashes (the getHash() method still returns an int64_t since that is what goes across the wire).

In The marshalling code, cast to uint64_t before bit-shifting