Fixes a number of bugs: 1) Fixes an issue where collapsed sections would cause sections below them to indent. 2) Fixes an issue where the target for clicking on sparklines was slightly off
Changes behavior for minimizing sections: reduces the click target from anywhere in the section that was not a sparkline to just the top bold text.  I think this is much more intuitive now, since there are other things to click, that are often clicked more than the section minimization.  This should reduce missed clicks causing sections to minimize.