Simplify lcm-java/Makefile.am
The important change is that this uses a *single target* for make to
unpack the jars for the extensions to jchart2d, then pack all the
classes into the single `lcm.jar`.

With the extractions as separate targets, `make` often ran both
extractions (i.e. `jar -xf` commands) at the same time using different
cores. This would cause a permissions issue when both jobs tried to
write to `build/META-INF`.

You could argue that the `META-INF` directories for the extensions
should be preserved separately, but I'm leaving that out of this commit.

The other changes are just using `find` to populate lists of files to
clean and/or include in `dist_noinst_JAVA`, instead of the hard-coded
lists that were there before. This removes a couple redundant entries,
and hopefully renders the Makefiles easier to maintain.

I've also added `$(JARFILE)` to `CLEANFILES` so that `make clean` will
remove the compiled and packaged `lcm.jar`.