Add option to not generate __init__.py
Add an option to lcm-gen to suppress generation of the per-package
__init__.py. This logic is problematic because it relies on file locking
(which is not implemented on Windows) if one instance of lcm-gen is
invoked per input file (which allows for parallelism and faster
incremental generation), as is done by our provided CMake helpers.

We still want an __init__.py, of course, but this logic will be moved to
the CMake helper, which already knows everything necessary to generate
it.