Update Java export functions
Replace our home-grown implementation for exporting JAR targets with an
updated version of UseJava.cmake that has been submitted for inclusion
in CMake 3.7. (Essentially, the logic from the previous commit has been
merged into UseJava.cmake and submitted for upstream inclusion. Since
3.7 won't be released for some time, and we likely won't want to depend
on it right away anyway, we are including local copies of the respective
modules until such time as we do switch to CMake 3.7 as a minimum
dependency.)