Fix Windows build
For some reason, LCM uses the non-standard `WIN32` as its platform check
macro (rather than the standard `_WIN32` that is defined by the
compiler). For now at least, rather than modify all of the sources,
modify the build to define the symbol (and also the symbol to shut up
the stupid "insecure" warnings).