Add command line utilities
Add the command line utilities lcm-spy and lcm-logplayer-gui. Since
CMake encourages relocatable installations, simply hard-coding the full
path to the jar is inadequate. Instead, try to determine the path of the
script (using perl if available; otherwise, invoking the tools via a
symlink in a different directory will fail) and the jar file (so the
scripts can be run in the build directory as well as in an install). As
a result, the scripts don't actually change depending on the build
environment, so the "final" scripts can be kept in the source directory
and just copied as needed.

Also, fix the name of the JAR (oops).