Fix ninja gripe
Rename targets to copy scripts and create additional "alias" targets in
the root CMakeLists.txt to the "real" targets. This fixes an issue where
ninja is unhappy because CMake implicitly names the actual target as the
stated name prefixes with the relative subdirectory, which happens to be
the same as the name of the file (relative to the build directory) being
created, resulting in conflicting rules. Renaming the "real" targets
avoids this, while the "alias" targets allows the user to still be able
to `make lcm-spy` (or `ninja lcm-spy`)... although with make, this will
only work from the build root, and not the lcm-java subdirectory.