Fix build on Windows (again)
The more recently added utility executables were not building on Windows
for various reasons. Add a new (static) Windows porting library for the
WinSpecific (getopt) code, and include this (and in one case, ws2_32) in
link libraries of executables that require it. Modify FindGLib2.cmake to
include ws2_32 and winmm as link libraries of glib. Fix some missing
export decorations for functions in the main library that may be
required by utility executables. Fix unguarded use of `extern "C"`.