Generate documentation with CMake
Add targets to the CMake build logic to generate the documentation. In
keeping with the previously existing script, these start by removing any
previously generated documentation, and are not built unless the target
is specifically named.

Note that this REPLACES the old script, meaning that generating the
documentation now requires that CMake is used. This is because CMake
writes the documentation to the build directory (as one would expect),
which requires configuring the Doxyfiles, and is thus incompatible with
the old, non-configured Doxyfiles.