Disable sanitization in lcmgen
Remove clang address/memory sanitization options when building lcmgen,
as it is known to be sloppy about memory management and (per the comment
in the source) we don't really care, since non-systemic leaks in a short
lived process like lcmgen aren't really a problem. However, building
with sanitization results in a broken executable due to the errors.

(Users might want to build with sanitization for the sake of the
libraries, or because they are building LCM as part of a "superbuild"
and have sanitization turned on for all subprojects.)