Avoid use of BASH_SOURCE in wrapper scripts
The wrapper scripts use !#/bin/sh, not #!/bin/bash, and we'd prefer to
avoid the latter in case of systems that don't have it, if possible.
Accordingly, replace use of ${BASH_SOURCE[0]} with plain old $0. These
should be the same except if the script is sourced via another script,
and it seems safe enough to assume that doesn't happen for these.