Fix finding Python when multiple versions present
Create a simple FindPython.cmake module to find both the Python
interpreter and libraries, using the install prefix of the former as a
hint for the latter. This should reduce the chances of finding a
mismatched interpreter and libraries, as can happen on some machines if
an interpreter from a non-standard prefix is in the user's PATH. It also
lets us simplify the logic to find "Python" by having one module that
provides both components.

This specifically was causing problems on some OS X machines using
Python from Homebrew.