Create header-only link target lcm-coretypes
Introduce a new exported INTERFACE library, lcm-coretypes, which
provides users a way to depend on the LCM headers without needing to
link to the actual LCM library. LCM types libraries will likely want to
"link" to this.

(In particular, C++ LCM types "libraries" are header-only with an
include dependency on LCM but no link dependency. Consumers of these
"libraries" need the LCM include directories, but may not need to link
to the actual LCM library. The interface library gives consumers a way
to make this work, conveniently.)