Make Python module monolithic
Build both a static and shared version of the main LCM library. Modify
the Python module to use the static library, so that the Python module
does not depend on the shared library. (This is more convenient if using
LCM in a non-standard path, as it avoids possibly needing to set
LD_LIBRARY_PATH or similar in addition to PYTHONPATH.)