Refactor version handling
Change LCM's version symbols to match the names used by CMake (the old
names are still provided for compatibility). Teach CMake and setup.py to
read the version information from lcm_version.h rather than duplicating
it. Update the Lua documentation.

This leaves only the Lua rock having a duplicate copy of the version
information. This is probably unavoidable if the Lua rock is to continue
to build stand-alone, especially as the version information is part of
the file name of the .rockspec file.

(Also, .NET would be included as having another copy of the version
information, but it seems to not have correct version information at all
at this time.)