Teach lcmgen to add export decorations
Add two new options to lcmgen for C bindings, to add a supplemental
`#include` of a header that defines an export decoration symbol, and to
use such symbol to decorate the API prototypes.

This finally makes it possible to build a shared C binding library on
Windows without having to hand-craft a .def file.