CMake improvements to fix race condition
Add additional new CMake helper functions for generating C and C++
bindings libraries. Notably, these take care of setting up dependencies
so that the binding sources are not compiled before all bindings in the
library have been generated, which is necessary for some generators (at
least Ninja) to avoid a race condition trying to compile an object for
an LCM type that consumes the headers for other types before such
headers have been generated. This does something similar for C++
bindings (on sufficiently new CMake) and provides a second helper to
consume C++ bindings (which adds the necessary dependency for older
CMake), and streamlines linking bindings libraries to the requisite LCM
components.

The test types are also updated to use these new features.