Tweaks to .gitignore
Modify .gitignore in lcm-python to enforce that the ignored build
directory a) is actually a directory, and b) is exactly where we expect
it. This will continue to ignore what we intend to ignore, but won't
ignore some things that would match the old rule but aren't what we
intend to ignore. Add an ignore for the .pyc generated from the test
client.py when running the unit tests.