Bump version to 1.4.0 pre-release (1.3.95)
Bump the version number, so that users can tell the difference between
v1.3.x (LTS) and what will eventually be 1.4.x. Since we can't readily
add trailer tags, and since they make version comparisons a pain, this
uses the convention of using a very large number for the patch version.