Aggressively disable sanitization in lcmgen
Remove everything that looks like a sanitizer flag when building
lcm-gen. This will catch other sanitizers, as well as supplemental
sanitizer related flags besides `-fsanitize=<type>`, which also need to
be stripped to prevent the compiler complaining about them when the
'main' flag has been stripped.

See also 6eb8668f3323 and the comment added to the CMakeLists.txt for
rationale.