Make lcm-gen C/C++ include guard span includes
Make the include guards in the C and C++ headers generated by lcm-gen
span all of the file, including the #include list, by moving up the
appropriate #ifndef/#define statements generated.

Having all of the header inside the include guards makes it manifest
that the header can be #include-d twice in client code without any side
effects. This helps both the reader of the file, and static analysis
tools that may warn about including a generated header twice, unless we
make it obvious that there is an include guard protecting the entire
header file.