Fix ThreadSanitizer issue
Move initialization of vtable structs to static initializers rather than
an _init function. This silences a tsan complaint about a data race
writing to memory that we might be concurrently reading. (In practice,
this is harmless, as the values written would be the values that are
already there, and anyway an unguarded write of pointer size or less is
usually safe by itself if there are no data dependencies, but tsan is
not being unreasonable to complain about it.)

At least for now, the change is only made for non-Windows platforms, as
MSVC has a bad track record of doing static initialization properly.

This partly reverts one of the changes from the 286985e3d0e2 merge.