In generated C++ code, use static_cast instead of C cast (#199)
* Remove an unnecessary C-style cast

The return type of getHash() is already int64_t.  Using a C-style cast
in C++ code triggers -Wold-style-cast in both GCC and Clang.

* In generated C++ code, use static_cast instead of C cast

Using a C-style cast in C++ code triggers -Wold-style-cast in both GCC
and Clang, and because lcm-cpp is header-only it is sometimes difficult
for downstream projects to suppress this warning in LCM-generated code.