Channel based Golang wrapper for the C LCM library
This patch introduces a Go idiomatic way of communicating over LCM.
Instead of using a handler function-based implementation (like C) it
uses a golang idiomatic channel based variant in which the data that is
received via LCM, is sent down a Golang channel of type []byte. The
publish method is also changed from functio-based to instead return a
channel on which the client sends outgoing messages.

Futhermore, an LCM object does not contain the subscriptions any longer.
Instead, a map that keeps track of the subscriptions was added. This
simplyfies the code and tremendously enhances execution time.

In summary Publish() is replaced by Publisher(), Subscribe() returns a
channel on which messages on the subscribed channel is forwarded.