Adds Go to examples, tutorial and readme
Complete new tutorial for Go which describes a simple way to get
started with the Go LCM library bindings. Description on how to use the
Go code genrator and a link to godoc. And a few examples of using LCM
in Go were added. Also updates the readme with Go specific info, and
adds the commiters of Go to the authors file.

(The link to godoc should work as soon as this patch is merged down to
lcm master.)

Co-Authored-By: Gustaf Johansson <gustaf@pinon.se>