mpudpm: fix lock sequence while parsing packets
When parsing UDP packets that are not valid LCM messages, the
mpudpm module's error handling would get the internal locks into
an invalid state. When attempting to read the next UDP packet, it
would abort() with the following message on some platforms:

Attempt to unlock mutex that was not locked

Fix this by re-acquiring the receive lock during error handling
so the next unlock at the top of the recv loop will not crash.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>