Tweak previous strcpy changes
For the changes to debug, there is no need to malloc there because 256
chars is more than enough to hold everything in DBG_NAMETAB. But if
there is garbage in that variable, make sure to null terminate.

For mpudpm and udpm, there is no need to copy past the end of the
string, so use strcpy for that. There is no need for strncpy because the
size is already checked in those functions.